+++ src/postorius/tests/test_migration.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 1998-2018 by the Free Software Foundation, Inc.
+# Copyright (C) 2018 by the Free Software Foundation, Inc.
 #
 # This file is part of Postorius.
 #
@@ -16,6 +16,17 @@
 # You should have received a copy of the GNU General Public License along with
 # Postorius.  If not, see <http://www.gnu.org/licenses/>.
 
+from io import StringIO
+from django.core.management import call_command
+from django.test import TestCase
 
-__version__ = '1.2.2'
-default_app_config = 'postorius.apps.PostoriusConfig'
+
+class TestMigrationPending(TestCase):
+
+    def test_pending_migration(self):
+        # Test that there aren't any pending migrations to be added.
+        output = StringIO()
+        # We will call the makemigrations command and check the output.
+        call_command(
+            'makemigrations', interactive=False, dry_run=True, stdout=output)
+        self.assertEqual(output.getvalue(), 'No changes detected\n')